+++ front/src/layouts/ExperimentConfiguration.jsx
@@ -23,33 +23,11 @@ import {
   StyledFloatingLabel,
 } from '../styles/globalComponents';
 
+import { getFullDefaultValues } from '../utils/values';
+
 const StyledContainer = styled(Container)`
   margin: 20px;
 `;
-
-async function getFullDefaultValues(parameterJsonSchema, choice = 'none') {
-  const { properties } = parameterJsonSchema;
-  if (typeof properties !== 'undefined') {
-    const parameters = Object.keys(properties);
-    const defaultValues = choice === 'none' ? {} : { choice };
-    parameters.forEach(async (param) => {
-      const val = properties[param].oneOf[0].default;
-      if (val !== undefined) {
-        defaultValues[param] = val;
-      } else {
-        const { parent } = properties[param].oneOf[0];
-        const fetchedOptions = await fetch(`${process.env.REACT_APP_GET_CHILDREN_ENDPOINT + parent}`);
-        const receivedOptions = await fetchedOptions.json();
-        const [first] = receivedOptions;
-        const fetchedParams = await fetch(`${process.env.REACT_APP_SELECT_MODEL_ENDPOINT + first}`);
-        const parameterSchema = await fetchedParams.json();
-        defaultValues[param] = await getFullDefaultValues(parameterSchema, first);
-      }
-    });
-    return (defaultValues);
-  }
-  return ({});
-}
 
 function AddModels({
   availableModels,
